+++ client_server/requirements.txt
@@ -1,3 +1,56 @@
+beautifulsoup4==4.9.3
+blis==0.7.4
+Bottleneck==1.3.2
+catalogue==2.0.4
+certifi==2021.5.30
+charset-normalizer==2.0.4
+click==7.1.2
+cycler==0.10.0
+cymem==2.0.5
+fastai==1.0.42
+fastcore==1.3.25
+fastprogress==1.0.0
+idna==3.2
+install==1.3.4
+Jinja2==3.0.1
+joblib==1.0.1
+kiwisolver==1.3.1
+MarkupSafe==2.0.1
+matplotlib==3.4.2
+murmurhash==1.0.5
+numexpr==2.7.3
+numpy==1.21.1
+nvidia-ml-py3==7.352.0
+opencv-python==4.5.3.56
+packaging==21.0
+pandas==1.3.1
+pathy==0.6.0
+Pillow==8.3.1
+preshed==3.0.5
+pydantic==1.8.2
+pyparsing==2.4.7
+python-dateutil==2.8.2
+pytz==2021.1
+PyYAML==5.4.1
+requests==2.26.0
+scikit-learn==0.24.2
+scipy==1.7.1
+six==1.16.0
+smart-open==5.1.0
+soupsieve==2.2.1
+spacy==3.1.1
+spacy-legacy==3.0.8
+srsly==2.4.1
+thinc==8.0.8
+threadpoolctl==2.2.0
+torch==1.7.1
+torchvision==0.1.8
+tqdm==4.62.0
+typer==0.3.2
+typing==3.7.4.3
+typing-extensions==3.10.0.0
+urllib3==1.26.6
+wasabi==0.8.2
 absl-py==0.9.0
 astor==0.8.1
 astunparse==1.6.3
