--- requirements.txt
@@ -1,105 +1,0 @@
-absl-py==0.9.0
-astor==0.8.1
-astunparse==1.6.3
-attrs==19.3.0
-backcall==0.1.0
-bleach==3.1.4
-cachetools==4.0.0
-certifi==2020.4.5.1
-chardet==3.0.4
-click==7.1.2
-cycler==0.10.0
-decorator==4.4.2
-defusedxml==0.6.0
-entrypoints==0.3
-Flask==1.1.2
-Flask-SocketIO==4.3.1
-gast==0.3.3
-google-auth==1.13.1
-google-auth-oauthlib==0.4.1
-google-pasta==0.2.0
-graphviz==0.13.2
-grpcio==1.28.1
-h5py==2.10.0
-idna==2.9
-importlib-metadata==1.6.0
-ipykernel==5.2.0
-ipython==7.13.0
-ipython-genutils==0.2.0
-ipywidgets==7.5.1
-itsdangerous==1.1.0
-jedi==0.16.0
-Jinja2==2.11.1
-joblib==0.14.1
-jsonschema==3.2.0
-jupyter==1.0.0
-jupyter-client==6.1.2
-jupyter-console==6.1.0
-jupyter-core==4.6.3
-Keras-Applications==1.0.8
-Keras-Preprocessing==1.1.2
-kiwisolver==1.2.0
-Markdown==3.2.1
-MarkupSafe==1.1.1
-matplotlib==3.2.1
-mistune==0.8.4
-natsort==7.1.0
-nbconvert==5.6.1
-nbformat==5.0.5
-notebook==6.0.3
-numpy==1.18.2
-oauthlib==3.1.0
-opencv-contrib-python==4.2.0.34
-opencv-python==4.2.0.34
-opt-einsum==3.2.0
-pandas==1.0.3
-pandocfilters==1.4.2
-parso==0.6.2
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==7.1.2
-prometheus-client==0.7.1
-prompt-toolkit==3.0.5
-protobuf==3.11.3
-ptyprocess==0.6.0
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pydot==1.4.1
-Pygments==2.6.1
-pyparsing==2.4.7
-pyrsistent==0.16.0
-pyshine==0.0.3
-python-dateutil==2.8.1
-python-engineio==3.13.2
-python-socketio==4.6.0
-pytz==2019.3
-pyzmq==19.0.0
-qtconsole==4.7.2
-QtPy==1.9.0
-requests==2.23.0
-requests-oauthlib==1.3.0
-rsa==4.0
-scikit-learn==0.22.2.post1
-scipy==1.4.1
-seaborn==0.10.0
-Send2Trash==1.5.0
-six==1.14.0
-tensorboard==2.3.0
-tensorboard-plugin-wit==1.7.0
-tensorflow==2.3.1
-tensorflow-addons==0.11.2
-tensorflow-estimator==2.3.0
-termcolor==1.1.0
-terminado==0.8.3
-testpath==0.4.4
-tornado==6.0.4
-tqdm==4.55.1
-traitlets==4.3.3
-typeguard==2.10.0
-urllib3==1.25.8
-wcwidth==0.1.9
-webencodings==0.5.1
-Werkzeug==1.0.1
-widgetsnbextension==3.5.1
-wrapt==1.12.1
-zipp==3.1.0
