--- src/models.py
@@ -1,58 +1,0 @@
-from tensorflow.keras.applications import MobileNetV2
-from tensorflow.keras.layers import Dense, Dropout, GlobalAveragePooling2D
-from tensorflow.keras.layers import TimeDistributed, GRU, BatchNormalization
-from tensorflow.keras.models import Sequential
-
-
-def create_model_wlasl20c(frames, width, height, channels, output):
-
-    model = Sequential([
-        # ConvNet
-        TimeDistributed(
-            MobileNetV2(weights='imagenet', include_top=False,
-                        input_shape=[height, width, channels]),
-            input_shape=[frames, height, width, channels]
-        ),
-        TimeDistributed(GlobalAveragePooling2D()),
-
-        # GRUs
-        GRU(256, return_sequences=True),
-        BatchNormalization(),
-        GRU(256),
-
-        # Feedforward
-        Dense(units=64, activation='relu'),
-        Dropout(0.65),
-        Dense(units=32, activation='relu'),
-        Dropout(0.65),
-        Dense(units=output, activation='softmax')
-    ])
-
-    return model
-
-
-
-
-def create_model_wlasl2000(frames, width, height, channels, output):
-
-    model = Sequential([
-        # ConvNet
-        TimeDistributed(
-            MobileNetV2(weights='imagenet', include_top=False,
-                        input_shape=[height, width, channels]),
-            input_shape=[frames, height, width, channels]
-        ),
-        TimeDistributed(GlobalAveragePooling2D()),
-
-        # GRUs
-        GRU(256, return_sequences=True),
-        BatchNormalization(),
-        GRU(256),
-
-        # Feedforward
-        Dense(units=1000, activation='relu'),
-        Dropout(0.75),
-        Dense(units=output, activation='softmax')
-    ])
-
-    return model
